requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cssselect==1.1.0
cssutils==2.4.2
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
lxml==4.9.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.9.0
platformdirs==2.5.2
premailer==3.10.0
pycparser==2.21
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
schemadict==0.0.12
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
whitenoise==6.2.0
yagmail==0.15.277
